related to #2586
Sending mails with the smtp_batch output might fail e.g. due to wrong
settings regarding the SMTP-Server. So far this is not reflected in the
exit code when running the bot via cli to trigger sending the mails.
Fixing this (at least for the non-interactive call) enables users to
integrate this into their monitoring properly.